added express routes and structured server on files
